Fix unrecognized type warning
the column `typinput` is of type `regproc` which is a special kind of
string that represents function names (see
https://www.postgresql.org/docs/current/datatype-oid.html).

But this type is not recognized by rails out of the box, and so it is
triggering a warning:

```
unknown OID 24: failed to recognize type of 'typinput'. It will be
treated as String.
```

Instead of letting Rails infer that it should treat that type as a
string, let's coerce it to a string directly in the query. This makes
the warning go away without any change in functionality.